open-webui-main/backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
pydantic==2.12.5
python-multipart==0.0.22
itsdangerous==2.2.0

python-socketio==5.16.1
python-jose==3.5.0
cryptography==46.0.5
bcrypt==5.0.0
argon2-cffi==25.1.0
PyJWT[crypto]==2.11.0
authlib==1.6.9

requests==2.32.5
aiohttp==3.13.2 # do not update to 3.13.3 - broken
async-timeout==5.0.1
aiocache==0.12.3
aiofiles==25.1.0
starlette-compress==1.7.0
Brotli==1.1.0
httpx[socks,http2,zstd,cli,brotli]==0.28.1
starsessions[redis]==2.2.1
python-mimeparse==2.0.0

sqlalchemy==2.0.48
alembic==1.18.4
peewee==3.19.0
peewee-migrate==1.14.3

pycrdt==0.12.47
redis==7.4.0

APScheduler==3.11.2
RestrictedPython==8.1
pytz==2026.1.post1

loguru==0.7.3
asgiref==3.11.1

# AI libraries
tiktoken==0.12.0
mcp==1.26.0

openai==2.29.0
anthropic==0.86.0
google-genai==1.66.0

langchain==1.2.10
langchain-community==0.4.1
langchain-classic==1.0.1
langchain-text-splitters==1.1.1

fake-useragent==2.2.0
chromadb==1.5.2
weaviate-client==4.20.3
opensearch-py==3.1.0

transformers==5.3.0
sentence-transformers==5.2.3
accelerate==1.13.0
pyarrow==20.0.0 # fix: pin pyarrow version to 20 for rpi compatibility #15897
einops==0.8.2 

ftfy==6.3.1
chardet==5.2.0
pypdf==6.7.5
fpdf2==2.8.7
pymdown-extensions==10.21
docx2txt==0.9
python-docx==1.1.2
python-pptx==1.0.2
reportlab==4.4.1
msoffcrypto-tool==6.0.0
unstructured==0.18.31

nltk==3.9.3
Markdown==3.10.2
beautifulsoup4==4.14.3
pypandoc==1.16.2
pandas==3.0.1
openpyxl==3.1.5
pyxlsb==1.0.10
xlrd==2.0.2
validators==0.35.0
psutil==7.2.2
sentencepiece==0.2.1
soundfile==0.13.1

pillow==12.1.1
opencv-python-headless==4.13.0.92
rapidocr-onnxruntime==1.4.4
rank-bm25==0.2.2

onnxruntime==1.24.3
faster-whisper==1.2.1

black==26.1.0
youtube-transcript-api==1.2.4
pytube==15.0.0

pydub==0.25.1
ddgs==9.11.3

azure-ai-documentintelligence==1.0.2
azure-identity==1.25.2
azure-storage-blob==12.28.0
azure-search-documents==11.6.0

## Google Drive
google-api-python-client==2.193.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0

googleapis-common-protos==1.72.0
google-cloud-storage==3.9.0

## Databases
pymongo==4.16.0
psycopg2-binary==2.9.11
pgvector==0.4.2

PyMySQL==1.1.2
boto3==1.42.62
# mariadb==1.1.14 should be added if you want to support MariaDB

pymilvus==2.6.9
qdrant-client==1.17.0
playwright==1.58.0 # Caution: version must match docker-compose.playwright.yaml - Update the docker-compose.yaml if necessary
elasticsearch==9.3.0
pinecone==6.0.2
oracledb==3.4.2

av==14.0.1 # Caution: Set due to FATAL FIPS SELFTEST FAILURE, see discussion https://github.com/open-webui/open-webui/discussions/15720

colbert-ai==0.2.22


## Tests
docker~=7.1.0
pytest~=8.4.1
pytest-docker~=3.2.5

## LDAP
ldap3==2.9.1

## Firecrawl
firecrawl-py==4.18.0

## Trace
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-fastapi==0.61b0
opentelemetry-instrumentation-sqlalchemy==0.61b0
opentelemetry-instrumentation-redis==0.61b0
opentelemetry-instrumentation-requests==0.61b0
opentelemetry-instrumentation-logging==0.61b0
opentelemetry-instrumentation-httpx==0.61b0
opentelemetry-instrumentation-aiohttp-client==0.61b0
opentelemetry-instrumentation-system-metrics==0.61b0
